+++ ds-test/src/main/java/org/vaadin/tori/data/TestDataSource.java
@@ -197,15 +197,15 @@ public class TestDataSource implements DataSource {
     }
 
     @Override
-    public int getThreadCount(final Category category)
+    public long getThreadCount(final Category category)
             throws DataSourceException {
-        return executeWithEntityManager(new Command<Integer>() {
+        return executeWithEntityManager(new Command<Long>() {
             @Override
-            public Integer execute(final EntityManager em) {
-                final TypedQuery<Integer> q = em
+            public Long execute(final EntityManager em) {
+                final TypedQuery<Long> q = em
                         .createQuery(
                                 "select count(t) from DiscussionThread t where t.category = :category",
-                                Integer.class);
+                                Long.class);
                 q.setParameter("category", category);
                 return q.getSingleResult();
             }
